refactor : `answers` 1급 컬렉션 타입 생성
