+++ src/main/java/qna/domain/answer/Answer.java
@@ -81,7 +81,7 @@ public class Answer extends BaseTimeEntity {
 
     public void toQuestion(Question question) {
         if (this.question != null) {
-            this.getQuestion().getAnswers().remove(this);
+            this.getQuestion().removeAnswer(this);
         }
         this.question = question;
         question.getAnswers().add(this);
